+++ packages/send/src/send-get-latest-block.js
@@ -1,7 +1,8 @@
 import {GetLatestBlockRequest, AccessAPI} from "@onflow/protobuf"
 import {response} from "@onflow/response"
-import {bufferToHexString} from "@onflow/bytes"
 import {unary} from "./unary"
+
+const u8ToHex = u8 => Buffer.from(u8).toString("hex")
 
 export async function sendGetLatestBlock(ix, opts = {}) {
   const req = new GetLatestBlockRequest()
@@ -18,17 +19,17 @@ export async function sendGetLatestBlock(ix, opts = {}) {
   const ret = response()
   ret.tag = ix.tag
   ret.latestBlock = {
-    id: bufferToHexString(latestBlock.getId_asU8()),
-    parentId: bufferToHexString(latestBlock.getParentId_asU8()),
+    id: u8ToHex(latestBlock.getId_asU8()),
+    parentId: u8ToHex(latestBlock.getParentId_asU8()),
     height: latestBlock.getHeight(),
     timestamp: latestBlock.getTimestamp(),
     collectionGuarantees: collectionGuarantees.map(collectionGuarantee => ({
-      collectionId: bufferToHexString(collectionGuarantee.getCollectionId_asU8()),
-      signatures: collectionGuarantee.getSignaturesList()
+      collectionId: u8ToHex(collectionGuarantee.getCollectionId_asU8()),
+      signatures: collectionGuarantee.getSignaturesList(),
     })),
     blockSeals: blockSeals.map(blockSeal => ({
-      blockId: bufferToHexString(blockSeal.getBlockId_asU8()),
-      executionReceiptId: bufferToHexString(blockSeal.getExecutionReceiptId_asU8()),
+      blockId: u8ToHex(blockSeal.getBlockId_asU8()),
+      executionReceiptId: u8ToHex(blockSeal.getExecutionReceiptId_asU8()),
       executionReceiptSignatures: blockSeal.getExecutionReceiptSignaturesList(),
       resultApprovalSignatures: blockSeal.getResultApprovalSignaturesList(),
     })),