+++ packages/send/src/send-get-transaction-status.js
@@ -1,11 +1,13 @@
 import {AccessAPI, GetTransactionRequest} from "@onflow/protobuf"
 import {response} from "@onflow/response"
-import {bufferToHexString, hashToBuffer} from "@onflow/bytes"
 import {unary} from "./unary"
+
+const u8ToHex = u8 => Buffer.from(u8).toString("hex")
+const hexBuffer = hex => Buffer.from(hex, "hex")
 
 export async function sendGetTransactionStatus(ix, opts = {}) {
   const req = new GetTransactionRequest()
-  req.setId(hashToBuffer(ix.txId))
+  req.setId(hashBuffer(ix.txId))
 
   const res = await unary(opts.node, AccessAPI.GetTransactionResult, req)
 
@@ -19,7 +21,7 @@ export async function sendGetTransactionStatus(ix, opts = {}) {
     errorMessage: res.getErrorMessage(),
     events: events.map(event => ({
       type: event.getType(),
-      transactionId: bufferToHexString(event.getTransactionId_asU8()),
+      transactionId: u8ToHex(event.getTransactionId_asU8()),
       transactionIndex: event.getTransactionIndex(),
       eventIndex: event.getEventIndex(),
       payload: event.getPayload_asU8(),